pyproject.toml

# -----------------------------------------------------------------------------------------------------------------------------------------------------------
#                                                     *****     Project dependencies     *****
# -----------------------------------------------------------------------------------------------------------------------------------------------------------


[project]
name = "appnews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "nltk>=3.9.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.1",
    "requests>=2.32.5",
    "spacy>=3.8.7",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.50.0",
    "textblob>=0.19.0",
    "uvicorn>=0.37.0",
]
